Completely rewrite linker argument handling
Now, place arguments in the "linkerargs" file
and they will be passed to clang as-is. (This
means that you need to specify -Xlinker if
applicable.)

Design note: The -reexport_library lines are
for dylibs I have already built with darwinbuild,
while the -l lines are for dylibs from the pd17.4
beta image. The latter will be replaced with the
former as time moves on.